requirements.txt
flask
Flask-WTF
passlib
sightengine
numpy 
scipy
scikit-learn
tensorflow
pillow
h5py
keras
